requirements.txt
# Core dependencies
streamlit==1.28.0
fastapi==0.104.0
uvicorn==0.24.0
python-dotenv==1.0.0

# RAG and LLM
langchain==0.0.330
openai==1.3.0
chromadb==0.4.18
tiktoken==0.5.1

# Document processing
pypdf==3.17.0
markdown==3.5.1
beautifulsoup4==4.12.2

# Data processing
pandas==2.2.0
numpy==1.26.0
pydantic==2.4.2

# Visualization
plotly==5.18.0

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0 